+++ astroqtpy/base.py
