+++ astroqtpy/quadnodes.py
@@ -1,4 +1,4 @@
-from .base_node import QuadNode
+from .base import QuadNode
 
 class NbodyNode(QuadNode):
     """Quadtree node for Nbody simulations.
@@ -16,13 +16,21 @@ class NbodyNode(QuadNode):
     def __init__(self, x_min, x_max, y_min, y_max, depth=0):
         """__init__
 
-        Create the N-body quad tree node for astroQTpy.
+        Create the N-body quadtree node for astroQTpy.
         """
         super().__init__(x_min, x_max, y_min, y_max, depth)
         
         
     def generate_node_value(self):
+        """_summary_
+
+        _extended_summary_
+        """
         pass
     
     def split_node(self):
+        """_summary_
+
+        _extended_summary_
+        """
         pass
