+++ astroqtpy/quadtree.py
@@ -1,0 +1,4 @@
+from quadnodes import NbodyNode
+
+
+
