--- astroqtpy/quatree.py
@@ -1,4 +1,0 @@
-from nodes import NbodyNode
-
-
-
