+++ docs/source/api.rst
@@ -6,5 +6,5 @@ Detailed API Documentation
 .. toctree::
    :titlesonly:
    
-   nodes
-   base_node
+   quadnodes
+   base
