+++ docs/source/base.rst
@@ -1,0 +1,9 @@
+.. _base:
+
+Base Components
+================
+
+Base `QuadNode` and `QuadPoint` classes for astroQTpy.
+
+.. automodule:: astroqtpy.base
+   :members:
