--- docs/source/base_node.rst
@@ -1,9 +1,0 @@
-.. _base_node:
-
-Base Node
-=====================
-
-Base component classes for astroQTpy.
-
-.. automodule:: astroqtpy.base_node
-   :members:
