+++ docs/source/quadnodes.rst
@@ -1,0 +1,9 @@
+.. _quadnodes:
+
+Quadtree Nodes
+==============
+
+Quadtree node classes to inherit the base `QuadNode` in astroQTpy.
+
+.. automodule:: astroqtpy.quadnodes
+   :members:
